enhancement/1431 - move Hold classes to navlib
