+++ src/assets/scripts/client/navigationLibrary/HoldModel.js
