+++ src/assets/scripts/client/navigationLibrary/ProcedureModel.js
@@ -5,7 +5,7 @@ import _map from 'lodash/map';
 import _random from 'lodash/random';
 import _uniq from 'lodash/uniq';
 import WaypointModel from '../aircraft/FlightManagementSystem/WaypointModel';
-import HoldCollection from '../airport/HoldCollection';
+import HoldCollection from './HoldCollection';
 import { PROCEDURE_TYPE } from '../constants/routeConstants';
 
 /**
